renamed src/index-*.js to src/bundle-.js
