+++ debug/segments.js
@@ -1,5 +1,5 @@
 // node --experimental-modules enumerate-segments.js
-import {Exifr} from '../src/index-full.js'
+import {Exifr} from '../src/bundle-full.js'
 import {promises as fs} from 'fs'
 import path from 'path'
