+++ src/bundle-core.js
