+++ test/highlevel-gps.spec.js
@@ -1,7 +1,7 @@
 import {assert} from './test-util.js'
 import {getFile, testSegment, testSegmentTranslation, testPickOrSkipTags} from './test-util.js'
-import {Exifr} from '../src/index-full.js'
-import * as exifr from '../src/index-full.js'
+import {Exifr} from '../src/bundle-full.js'
+import * as exifr from '../src/bundle-full.js'
 import {gpsOnlyOptions} from '../src/options.js'
