+++ test/reader.spec.js
@@ -1,7 +1,7 @@
 import {promises as fs} from 'fs'
 import {assert} from './test-util.js'
 import {isBrowser, isNode, getPath, getFile} from './test-util.js'
-import * as exifr from '../src/index-full.js'
+import * as exifr from '../src/bundle-full.js'
 
 
 export function createImg(url) {
